refactor: consolidate error state component [skip version]
Preserve ErrorState SDK compatibility and accessibility while using one canonical implementation.
