+++ frontend/src/components/ErrorBoundary.jsx
@@ -1,7 +1,7 @@
 import { Component } from 'react';
-import { AlertTriangle, RefreshCw } from 'lucide-react';
 import api from '../services/api';
 import { t } from '../i18n/t';
+import ErrorState from './ErrorState';
 
 // Class-based Error Boundary for catching React errors
 export class ErrorBoundary extends Component {
@@ -53,68 +53,16 @@ export class ErrorBoundary extends Component {
     render() {
         if (this.state.hasError) {
             return (
-                <div className="sk-error-state" role="alert">
-                    <div className="sk-error-state__icon">
-                        <AlertTriangle size={32} />
-                    </div>
-                    <h3 className="sk-error-state__title">
-                        {t('common.error.title', 'Something went wrong')}
-                    </h3>
-                    <p className="sk-error-state__message">
-                        {t('common.error.unexpected', 'An unexpected error occurred')}
-                    </p>
-                    <button type="button" className="btn btn-primary" onClick={this.handleRetry}>
-                        <RefreshCw size={14} />
-                        {t('common.actions.tryAgain', 'Try again')}
-                    </button>
-                </div>
+                <ErrorState
+                    title={t('common.error.title', 'Something went wrong')}
+                    message={t('common.error.unexpected', 'An unexpected error occurred')}
+                    onRetry={this.handleRetry}
+                />
             );
         }
 
         return this.props.children;
     }
-}
-
-// Functional component for displaying API/fetch errors
-export function ErrorState({
-    title = t('common.error.failedToLoad', 'Failed to load'),
-    message,
-    error,
-    onRetry,
-    compact = false
-}) {
-    const errorMessage = message || error?.message
-        || t('common.error.loadingData', 'An error occurred while loading data');
-
-    if (compact) {
-        return (
-            <div className="error-state-compact">
-                <AlertTriangle size={16} />
-                <span>{errorMessage}</span>
-                {onRetry && (
-                    <button type="button" className="btn btn-ghost btn-sm" onClick={onRetry}>
-                        <RefreshCw size={14} /> {t('common.actions.retry', 'Retry')}
-                    </button>
-                )}
-            </div>
-        );
-    }
-
-    return (
-        <div className="sk-error-state">
-            <div className="sk-error-state__icon">
-                <AlertTriangle size={32} />
-            </div>
-            <h3 className="sk-error-state__title">{title}</h3>
-            <p className="sk-error-state__message">{errorMessage}</p>
-            {onRetry && (
-                <button type="button" className="btn btn-primary" onClick={onRetry}>
-                    <RefreshCw size={14} />
-                    {t('common.actions.tryAgain', 'Try again')}
-                </button>
-            )}
-        </div>
-    );
 }
 
 export default ErrorBoundary;
